requirements.txt
ultralytics>=8.1.0,<9.0
roboflow>=1.1.0,<2.0
python-dotenv>=1.0.0,<2.0
pyyaml>=6.0,<7.0
opencv-python>=4.8.0,<5.0
numpy>=1.23.0,<2.0
